backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI/ML Libraries
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
sentence-transformers==2.2.2
chromadb==0.4.18
transformers==4.36.2
torch==2.1.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# CORS
fastapi-cors==0.0.6
